Add a progress indicator header to Non-TTY progress bars
This adds a progress indicator header of the form:

```
0%      50%      100%
|--------|---------|
....
```

to the Non-TTY progress bar, so that progress can be more clearly visualized in
settings where there is no opportunity to rewind, and have an 'interactive'
progressbar.

For comparison, the old setup would yield:

```
....
```

In the same scenario as outlined above.

Changelog: Title
Signed-off-by: Ole Petter <ole.orhagen@northern.tech>